+++ Basic/Combination.h
@@ -5,12 +5,12 @@
 
 class Combination : std::exception {
 private:
-	using Vector = std::vector< unsigned int >;
+	using Vector = std::vector< BigInteger >;
 	using Matrix = std::vector< Vector >;
 public:
 	void CreatePascalTable();//파스칼 테이블 제작
 	void PrintPascalTable() const;//
-	unsigned int N_Combination_M(const size_t& n,const size_t& m) const;//nCm
+	BigInteger N_Combination_M(const size_t& n,const size_t& m) const;//nCm
 public://Const, Dest, getter and setter
 	Combination(const size_t& size) :size_(size) {
 		set_size(size_);
@@ -25,7 +25,7 @@ public://Const, Dest, getter and setter
 	size_t get_size() const { return size_; }
 	
 private:
-	unsigned int PascalUility(const size_t& n, const size_t& m);
+	BigInteger PascalUility(const size_t& n, const size_t& m);
 private:
 	Matrix pascal_table_;
 	size_t size_;//nCm에서 n
